Move hand type to its own files
