+++ src/main/kotlin/hwr/oop/group1/poker/HandRank.kt
@@ -1,18 +1,5 @@
 package hwr.oop.group1.poker
 
-enum class HandType(val strength: Int) {
-    HIGH_CARD(1),
-    ONE_PAIR(2),
-    TWO_PAIR(3),
-    THREE_OF_A_KIND(4),
-    STRAIGHT(5),
-    FLUSH(6),
-    FULL_HOUSE(7),
-    FOUR_OF_A_KIND(8),
-    STRAIGHT_FLUSH(9),
-    ROYAL_FLUSH(10)
-}
-
 data class HandRank(
     val type: HandType,
     val ranks: List<CardRank> // ordered for tie-breaker
