+++ src/main/kotlin/hwr/oop/group1/poker/HandType.kt
@@ -1,0 +1,14 @@
+package hwr.oop.group1.poker
+
+enum class HandType(val strength: Int) {
+    HIGH_CARD(1),
+    ONE_PAIR(2),
+    TWO_PAIR(3),
+    THREE_OF_A_KIND(4),
+    STRAIGHT(5),
+    FLUSH(6),
+    FULL_HOUSE(7),
+    FOUR_OF_A_KIND(8),
+    STRAIGHT_FLUSH(9),
+    ROYAL_FLUSH(10)
+}
